+++ simulation/dtos/department_dtos.py
@@ -4,13 +4,13 @@ from modules.system.api import CurrencyCode # Added for Phase 33
 
 @dataclass(frozen=True)
 class FinanceStateDTO:
-    balance: Dict[CurrencyCode, float] # Changed for Phase 33
-    revenue_this_turn: Dict[CurrencyCode, float] # Changed for Phase 33
-    expenses_this_tick: Dict[CurrencyCode, float] # Changed for Phase 33
+    balance: Dict[CurrencyCode, int] # Changed to int (pennies)
+    revenue_this_turn: Dict[CurrencyCode, int] # Changed to int (pennies)
+    expenses_this_tick: Dict[CurrencyCode, int] # Changed to int (pennies)
     consecutive_loss_turns: int
-    profit_history: List[float] # This might need to be Dict if it's total profit in a specific currency, but for now let's keep it simple or change to Dict
+    profit_history: List[int] # Changed to int (pennies)
     altman_z_score: float
-    valuation: Dict[CurrencyCode, float] # Changed for Phase 33
+    valuation: Dict[CurrencyCode, int] # Changed to int (pennies)
     total_shares: float
     treasury_shares: float
     dividend_rate: float
@@ -21,7 +21,7 @@ class ProductionStateDTO:
     current_production: float
     productivity_factor: float
     production_target: float
-    capital_stock: float
+    capital_stock: int # Changed to int (pennies)
     base_quality: float
     automation_level: float
     specialization: str
@@ -32,10 +32,10 @@ class ProductionStateDTO:
 @dataclass(frozen=True)
 class SalesStateDTO:
     inventory_last_sale_tick: Dict[str, int]
-    price_history: Dict[str, float]
+    price_history: Dict[str, int] # Changed to int (pennies)
     brand_awareness: float
     perceived_quality: float
-    marketing_budget: float
+    marketing_budget: int # Changed to int (pennies)
 
 @dataclass(frozen=True)
 class HRStateDTO:
